backend/pyproject.toml
[tool.poetry]
name = "strava-insights-backend"
version = "0.1.0"
description = "Backend API for Strava Insights"
authors = ["Strava Insights"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.13"
fastapi = "0.116.1"
uvicorn = {version = "0.35.0", extras = ["standard"]}
sqlalchemy = "2.0.43"
psycopg = {version = "3.2.9", extras = ["binary"]}
redis = "5.2.1"
pydantic-settings = "2.10.1"
alembic = "1.16.5"
celery = {version = "5.5.3", extras = ["redis"]}
cryptography = "45.0.7"
itsdangerous = "2.2.0"
httpx = "0.28.1"

[tool.poetry.group.dev.dependencies]
pytest = "8.4.1"

[build-system]
requires = ["poetry-core>=2.0.0"]
build-backend = "poetry.core.masonry.api"
